+++ CHANGES.md
@@ -1,4 +1,14 @@
 # Changelog #
 
+* @dev ()
+  * String format for commented-out entries can now be configured.
+  * Added option to specify comment character.
+  * Added option to specify key-value separator character.
+  * Missing translation file no longer aborts.
+  * Fixed handling `!` as comment line marker.
+  * You can use either `prop-tool` or `proptool` as tool name.
+  * Reference file is now also checked for common errors.
+  * Trailing chars check can also detect missing "\n" literal too.
+
 * v1.0.0 (2021-07-16)
-    * Initial public release
+  * Initial public release.